refactor/folder structure refactoring
